Added a script interface to Data (QData) now Data implements QData; Added a Scriptviewer, with a script editing textBox. changed the size_t's to unsigned ints
git-svn-id: https://tulip-plugins.googlecode.com/svn/trunk@115 4ddaf586-9532-11dd-8e9d-3bce277bc375

+++ miniprojet/IData.h
@@ -23,8 +23,8 @@ class IData : public Observable
 		virtual void add(T, T) = 0;
 		virtual T getX(int) const = 0;
 		virtual T getY(int) const = 0;
-		virtual void setX(size_t, T) = 0;
-		virtual void setY(size_t, T) = 0;
+		virtual void setX(unsigned int, T) = 0;
+		virtual void setY(unsigned int, T) = 0;
 		virtual T getXMin() const = 0;
 		virtual T getYMin() const = 0;
 		virtual T getXMax() const = 0;
